Generate typos using multiple methods

This is a simple wrapper to let users generate typos using multiple methods in a single function call. Results returned by this function should be equivalent to the union of all functions supplied to methods.

generate_typos(vec, methods, chars = NA)

Arguments

vec
A vector of strings to attempt to generate typos for.

methods
A vector of strings matching the methods (listed in typo_generation) to use for typo generation.

chars
Characters to use in generating typos. If set to NA, will use the default character sets for each method; otherwise, chars[[i]] should be a character set usable by methods[[i]].

Value

A named list of length length(vec) containing all unique permutations from the typo-generating methods.

Examples

generate_typos(c("Michael", "David"), c("typo_addition", "typo_prefix"))

#> $Michael
#>  [1] "Michael-" "Michael." "Michael0" "Michael1" "Michael2" "Michael3"
#>  [7] "Michael4" "Michael5" "Michael6" "Michael7" "Michael8" "Michael9"
#> [13] "Michael_" "Michaela" "Michaelb" "Michaelc" "Michaeld" "Michaele"
#> [19] "Michaelf" "Michaelg" "Michaelh" "Michaeli" "Michaelj" "Michaelk"
#> [25] "Michaell" "Michaelm" "Michaeln" "Michaelo" "Michaelp" "Michaelq"
#> [31] "Michaelr" "Michaels" "Michaelt" "Michaelu" "Michaelv" "Michaelw"
#> [37] "Michaelx" "Michaely" "Michaelz" "-Michael" ".Michael" "0Michael"
#> [43] "1Michael" "2Michael" "3Michael" "4Michael" "5Michael" "6Michael"
#> [49] "7Michael" "8Michael" "9Michael" "_Michael" "aMichael" "bMichael"
#> [55] "cMichael" "dMichael" "eMichael" "fMichael" "gMichael" "hMichael"
#> [61] "iMichael" "jMichael" "kMichael" "lMichael" "mMichael" "nMichael"
#> [67] "oMichael" "pMichael" "qMichael" "rMichael" "sMichael" "tMichael"
#> [73] "uMichael" "vMichael" "wMichael" "xMichael" "yMichael" "zMichael"
#> 
#> $David
#>  [1] "David-" "David." "David0" "David1" "David2" "David3" "David4" "David5"
#>  [9] "David6" "David7" "David8" "David9" "David_" "Davida" "Davidb" "Davidc"
#> [17] "Davidd" "Davide" "Davidf" "Davidg" "Davidh" "Davidi" "Davidj" "Davidk"
#> [25] "Davidl" "Davidm" "Davidn" "Davido" "Davidp" "Davidq" "Davidr" "Davids"
#> [33] "Davidt" "Davidu" "Davidv" "Davidw" "Davidx" "Davidy" "Davidz" "-David"
#> [41] ".David" "0David" "1David" "2David" "3David" "4David" "5David" "6David"
#> [49] "7David" "8David" "9David" "_David" "aDavid" "bDavid" "cDavid" "dDavid"
#> [57] "eDavid" "fDavid" "gDavid" "hDavid" "iDavid" "jDavid" "kDavid" "lDavid"
#> [65] "mDavid" "nDavid" "oDavid" "pDavid" "qDavid" "rDavid" "sDavid" "tDavid"
#> [73] "uDavid" "vDavid" "wDavid" "xDavid" "yDavid" "zDavid"
#>