Create a list of characters which neighbor one another on QWERTY keyboards

Create a list of characters which neighbor one another on QWERTY keyboards

qwerty_neighbors()

Value

A list of length 37. Each element of the list is named with a key on the QWERTY keyboard and contains a character vector of neighboring keys.

Examples

qwerty_neighbors()

#> $q
#> [1] "1" "2" "w" "a"
#> 
#> $w
#> [1] "2" "3" "e" "s" "a" "q"
#> 
#> $e
#> [1] "3" "4" "r" "d" "s" "w"
#> 
#> $r
#> [1] "4" "5" "t" "f" "d" "e"
#> 
#> $t
#> [1] "5" "6" "y" "g" "f" "r"
#> 
#> $y
#> [1] "6" "7" "u" "h" "g" "t"
#> 
#> $u
#> [1] "7" "8" "i" "j" "h" "y"
#> 
#> $i
#> [1] "8" "9" "o" "k" "j" "u"
#> 
#> $o
#> [1] "9" "0" "p" "l" "k" "i"
#> 
#> $p
#> [1] "0" "-" "l" "o"
#> 
#> $a
#> [1] "q" "w" "s" "z"
#> 
#> $s
#> [1] "w" "e" "d" "x" "z" "a"
#> 
#> $d
#> [1] "e" "r" "f" "c" "x" "s"
#> 
#> $f
#> [1] "r" "t" "g" "v" "c" "d"
#> 
#> $g
#> [1] "t" "y" "h" "b" "v" "f"
#> 
#> $h
#> [1] "y" "u" "j" "n" "b" "g"
#> 
#> $j
#> [1] "u" "i" "k" "m" "n" "h"
#> 
#> $k
#> [1] "i" "o" "l" "m" "j"
#> 
#> $l
#> [1] "o" "p" "k"
#> 
#> $z
#> [1] "a" "s" "x"
#> 
#> $x
#> [1] "s" "d" "c" "z"
#> 
#> $c
#> [1] "d" "f" "v" "x"
#> 
#> $v
#> [1] "f" "g" "b" "c"
#> 
#> $b
#> [1] "g" "h" "n" "v"
#> 
#> $n
#> [1] "h" "j" "m" "b"
#> 
#> $m
#> [1] "j" "k" "n"
#> 
#> $`1`
#> [1] "2" "q"
#> 
#> $`2`
#> [1] "1" "3" "w" "q"
#> 
#> $`3`
#> [1] "2" "4" "e" "w"
#> 
#> $`4`
#> [1] "3" "5" "r" "e"
#> 
#> $`5`
#> [1] "4" "6" "t" "r"
#> 
#> $`6`
#> [1] "5" "7" "y" "t"
#> 
#> $`7`
#> [1] "6" "8" "u" "y"
#> 
#> $`8`
#> [1] "7" "9" "i" "u"
#> 
#> $`9`
#> [1] "8" "0" "o" "i"
#> 
#> $`0`
#> [1] "9" "p" "o" "-"
#> 
#> $`-`
#> [1] "0" "p"
#>