Methods for typo generation

These functions implement various strategies to generate variants of strings. This is a partial implementation of Normand (2020); some variants are not implemented due to being out of scope (such as homoglyph typos, which this package does not intend to exhaustively cover). All typo permutations are returned in lowercase, and case differences are not considered typos.

typo_addition(vec, chars = gh_allowed())

typo_prefix(vec, chars = gh_allowed())

typo_bitsquat(vec, chars = gh_allowed())

typo_repetition(vec, chars = NULL)

typo_insertion(vec, chars = gh_allowed())

typo_omission(vec, chars = NULL)

typo_vowelswap(vec, chars = c("a", "e", "i", "o", "u"))

typo_transposition(vec, chars = NULL)

typo_replace(vec, chars = qwerty_neighbors())

typo_doublehit(vec, chars = qwerty_neighbors())

typo_hyphenation(vec, chars = "-")

typo_subdomain(vec, chars = ".")

Arguments

vec
A vector of strings to attempt to generate typos for.

chars
Characters to use in generating typos. Takes one of three formats:

For typo_doublehit and typo_replace, a named list (such as the one returned by qwerty_neighbors) where the names of the list represent individual characters and the elements of the list are characters to append or replace.

For typo_addition, typo_prefix, typo_bitsquat, typo_insertion, and typo_vowelswap, a vector of acceptable characters to add, swap, or insert.

For typo_repetition, typo_omission, and typo_transposition, this argument is ignored.

Value

A list of length length(vec) using vec as names, with each element of the list containing a character vector of potential typos.

Methods Implemented

At the moment 12 methods are implemented.

typo_addition adds characters to the end of a string.

typo_prefix adds characters to the start of a string.

typo_bitsquat changes a single character at a time, moving through the length of the string (so that each permutation is one bit different from the original input).

typo_repetition repeats a single character to simulate pressing a key twice.

typo_insertion inserts a single character into each position in the string. typo_hyphenation is equivalent to typo_insertion(vec, chars = "-") and typo_subdomain is equivalent to typo_insertion(vec, chars = ".").

typo_omission removes a single character from the string.

typo_vowelswap replaces the vowels in the string with the other vowels. Values provided to chars are used to define the full set of vowels; by default, y is not considered to be a vowel.

typo_transposition swaps adjacent characters throughout the string.

typo_replace replaces a single character with a set of characters (such as those close to the key on a keyboard).

typo_doublehit is similar to typo_replace but inserts the wrong character rather than replacing with it (to simulate hitting two keys at once).

References

Thibault Normand (2020). Typogenerator: a typosquatting generator in Golang. Golang package version 0.2.0. https://github.com/zntrio/typogenerator

Examples

typo_bitsquat("Michael")

#> $Michael
#>   [1] "-ichael" ".ichael" "0ichael" "1ichael" "2ichael" "3ichael" "4ichael"
#>   [8] "5ichael" "6ichael" "7ichael" "8ichael" "9ichael" "_ichael" "aichael"
#>  [15] "bichael" "cichael" "dichael" "eichael" "fichael" "gichael" "hichael"
#>  [22] "iichael" "jichael" "kichael" "lichael" "michael" "nichael" "oichael"
#>  [29] "pichael" "qichael" "richael" "sichael" "tichael" "uichael" "vichael"
#>  [36] "wichael" "xichael" "yichael" "zichael" "m-chael" "m.chael" "m0chael"
#>  [43] "m1chael" "m2chael" "m3chael" "m4chael" "m5chael" "m6chael" "m7chael"
#>  [50] "m8chael" "m9chael" "m_chael" "machael" "mbchael" "mcchael" "mdchael"
#>  [57] "mechael" "mfchael" "mgchael" "mhchael" "mjchael" "mkchael" "mlchael"
#>  [64] "mmchael" "mnchael" "mochael" "mpchael" "mqchael" "mrchael" "mschael"
#>  [71] "mtchael" "muchael" "mvchael" "mwchael" "mxchael" "mychael" "mzchael"
#>  [78] "mi-hael" "mi.hael" "mi0hael" "mi1hael" "mi2hael" "mi3hael" "mi4hael"
#>  [85] "mi5hael" "mi6hael" "mi7hael" "mi8hael" "mi9hael" "mi_hael" "miahael"
#>  [92] "mibhael" "midhael" "miehael" "mifhael" "mighael" "mihhael" "miihael"
#>  [99] "mijhael" "mikhael" "milhael" "mimhael" "minhael" "miohael" "miphael"
#> [106] "miqhael" "mirhael" "mishael" "mithael" "miuhael" "mivhael" "miwhael"
#> [113] "mixhael" "miyhael" "mizhael" "mic-ael" "mic.ael" "mic0ael" "mic1ael"
#> [120] "mic2ael" "mic3ael" "mic4ael" "mic5ael" "mic6ael" "mic7ael" "mic8ael"
#> [127] "mic9ael" "mic_ael" "micaael" "micbael" "miccael" "micdael" "miceael"
#> [134] "micfael" "micgael" "miciael" "micjael" "mickael" "miclael" "micmael"
#> [141] "micnael" "micoael" "micpael" "micqael" "micrael" "micsael" "mictael"
#> [148] "micuael" "micvael" "micwael" "micxael" "micyael" "miczael" "mich-el"
#> [155] "mich.el" "mich0el" "mich1el" "mich2el" "mich3el" "mich4el" "mich5el"
#> [162] "mich6el" "mich7el" "mich8el" "mich9el" "mich_el" "michbel" "michcel"
#> [169] "michdel" "micheel" "michfel" "michgel" "michhel" "michiel" "michjel"
#> [176] "michkel" "michlel" "michmel" "michnel" "michoel" "michpel" "michqel"
#> [183] "michrel" "michsel" "michtel" "michuel" "michvel" "michwel" "michxel"
#> [190] "michyel" "michzel" "micha-l" "micha.l" "micha0l" "micha1l" "micha2l"
#> [197] "micha3l" "micha4l" "micha5l" "micha6l" "micha7l" "micha8l" "micha9l"
#> [204] "micha_l" "michaal" "michabl" "michacl" "michadl" "michafl" "michagl"
#> [211] "michahl" "michail" "michajl" "michakl" "michall" "michaml" "michanl"
#> [218] "michaol" "michapl" "michaql" "micharl" "michasl" "michatl" "michaul"
#> [225] "michavl" "michawl" "michaxl" "michayl" "michazl" "michae-" "michae."
#> [232] "michae0" "michae1" "michae2" "michae3" "michae4" "michae5" "michae6"
#> [239] "michae7" "michae8" "michae9" "michae_" "michaea" "michaeb" "michaec"
#> [246] "michaed" "michaee" "michaef" "michaeg" "michaeh" "michaei" "michaej"
#> [253] "michaek" "michaem" "michaen" "michaeo" "michaep" "michaeq" "michaer"
#> [260] "michaes" "michaet" "michaeu" "michaev" "michaew" "michaex" "michaey"
#> [267] "michaez"
#> 

typo_replace(c("Michael", "David"))

#> $Michael
#>  [1] "jichael" "kichael" "nichael" "m8chael" "m9chael" "mochael" "mkchael"
#>  [8] "mjchael" "muchael" "midhael" "mifhael" "mivhael" "mixhael" "micyael"
#> [15] "micuael" "micjael" "micnael" "micbael" "micgael" "michqel" "michwel"
#> [22] "michsel" "michzel" "micha3l" "micha4l" "micharl" "michadl" "michasl"
#> [29] "michawl" "michaeo" "michaep" "michaek"
#> 
#> $David
#>  [1] "eavid" "ravid" "favid" "cavid" "xavid" "savid" "dqvid" "dwvid" "dsvid"
#> [10] "dzvid" "dafid" "dagid" "dabid" "dacid" "dav8d" "dav9d" "davod" "davkd"
#> [19] "davjd" "davud" "davie" "davir" "davif" "davic" "davix" "davis"
#>