Generate typos using multiple methods

This is a simple wrapper to let users generate typos using multiple methods in a single function call. Results returned by this function should be equivalent to the union of all functions supplied to methods.

generate_typos(vec, methods, chars = NA)

Arguments

vec
A vector of strings to attempt to generate typos for.

methods
A vector of strings matching the methods (listed in typo_addition) to use for typo generation. Set to "*" or "all" to generate the maximum number of typos using default character sets.

chars
Characters to use in generating typos. If set to NA, will use the default character sets for each method; otherwise, chars[[i]] should be a character set usable by methods[[i]].

Value

A named list of length length(vec) containing all unique permutations from the typo-generating methods.

Examples

generate_typos(c("Michael", "David"), c("typo_addition", "typo_prefix"))

#> $Michael
#>  [1] "michael-" "michael." "michael0" "michael1" "michael2" "michael3"
#>  [7] "michael4" "michael5" "michael6" "michael7" "michael8" "michael9"
#> [13] "michael_" "michaela" "michaelb" "michaelc" "michaeld" "michaele"
#> [19] "michaelf" "michaelg" "michaelh" "michaeli" "michaelj" "michaelk"
#> [25] "michaell" "michaelm" "michaeln" "michaelo" "michaelp" "michaelq"
#> [31] "michaelr" "michaels" "michaelt" "michaelu" "michaelv" "michaelw"
#> [37] "michaelx" "michaely" "michaelz" "-michael" ".michael" "0michael"
#> [43] "1michael" "2michael" "3michael" "4michael" "5michael" "6michael"
#> [49] "7michael" "8michael" "9michael" "_michael" "amichael" "bmichael"
#> [55] "cmichael" "dmichael" "emichael" "fmichael" "gmichael" "hmichael"
#> [61] "imichael" "jmichael" "kmichael" "lmichael" "mmichael" "nmichael"
#> [67] "omichael" "pmichael" "qmichael" "rmichael" "smichael" "tmichael"
#> [73] "umichael" "vmichael" "wmichael" "xmichael" "ymichael" "zmichael"
#> 
#> $David
#>  [1] "david-" "david." "david0" "david1" "david2" "david3" "david4" "david5"
#>  [9] "david6" "david7" "david8" "david9" "david_" "davida" "davidb" "davidc"
#> [17] "davidd" "davide" "davidf" "davidg" "davidh" "davidi" "davidj" "davidk"
#> [25] "davidl" "davidm" "davidn" "davido" "davidp" "davidq" "davidr" "davids"
#> [33] "davidt" "davidu" "davidv" "davidw" "davidx" "davidy" "davidz" "-david"
#> [41] ".david" "0david" "1david" "2david" "3david" "4david" "5david" "6david"
#> [49] "7david" "8david" "9david" "_david" "adavid" "bdavid" "cdavid" "ddavid"
#> [57] "edavid" "fdavid" "gdavid" "hdavid" "idavid" "jdavid" "kdavid" "ldavid"
#> [65] "mdavid" "ndavid" "odavid" "pdavid" "qdavid" "rdavid" "sdavid" "tdavid"
#> [73] "udavid" "vdavid" "wdavid" "xdavid" "ydavid" "zdavid"
#>